[fix] Set tail and size of allocation log to uint64_t to stop overflows

+++ lib/allocator/device_structures.h
@@ -71,8 +71,8 @@ struct pr_region_operation_entry {
 
 struct pr_region_allocation_log {
 	uint64_t head_dev_offt;
-	uint32_t tail_dev_offt;
-	uint32_t size;
+	uint64_t tail_dev_offt;
+	uint64_t size;
 	uint64_t txn_id;
 } __attribute__((packed, aligned(32)));
